+++ NEWS.md
@@ -4,3 +4,10 @@
 
 * Remove the original make_PD function in lib.cpp.
 * Use Matrix::nearPD instead.
+
+# HiGarrote 1.0.2
+
+## Bug fixes
+
+* Remove the original is_PD function in lib.cpp.
+* Use matrixcalc::is.positive.definite instead.